# dj-connect-back/storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'DJ Connect API'
  description: 'API documentation for DJ Connect. All requests must include header Telegram-Init-Data: ${TWA.initData()}'
  contact:
    name: 'Contact Developer'
    url: 'https://t.me/beilec'
  license:
    name: 'Contact Lead'
    url: 'https://t.me/alievdenis1'
  version: 1.0.0
servers:
  -
    url: 'http://localhost:8082/api/v1'
    description: 'DJ Connect API Server'
  -
    url: 'https://dj-connect.xyz/api/v1'
paths:
  /dj/register:
    post:
      tags:
        - DJ
      summary: 'Register a new DJ'
      description: 'Registers a new DJ and returns the DJ object'
      operationId: registerDJ
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                stage_name:
                  type: string
                  example: 'DJ Example'
                city:
                  type: string
                  example: 'New York'
                payment_details:
                  type: string
                  example: 'Bank details or any payment information'
              type: object
      responses:
        200:
          description: 'Successful registration'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DJ'
        400:
          description: 'Bad Request'
        401:
          description: Unauthorized
        403:
          description: Forbidden
      security:
        -
          telegramAuth: []
  '/dj/profile/{id}':
    get:
      tags:
        - DJ
      summary: 'Get DJ profile'
      description: 'Returns DJ profile by ID'
      operationId: getDJProfile
      parameters:
        -
          name: id
          in: path
          description: 'DJ ID'
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 'Successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DJ'
        400:
          description: 'Bad Request'
        401:
          description: Unauthorized
        403:
          description: Forbidden
        404:
          description: 'Not Found'
      security:
        -
          telegramAuth: []
    put:
      tags:
        - DJ
      summary: 'Update DJ profile'
      description: 'Updates DJ profile information'
      operationId: updateDJProfile
      parameters:
        -
          name: id
          in: path
          description: 'DJ ID'
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                stage_name:
                  type: string
                  example: 'DJ Example'
                city:
                  type: string
                  example: 'New York'
                payment_details:
                  type: string
                  example: 'Bank details or any payment information'
              type: object
      responses:
        200:
          description: 'Successful update'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DJ'
        400:
          description: 'Bad Request'
        401:
          description: Unauthorized
        403:
          description: Forbidden
        404:
          description: 'Not Found'
      security:
        -
          telegramAuth: []
  '/dj/{dj_id}/track':
    post:
      tags:
        - DJ
      summary: 'Add a new track'
      description: 'Allows a DJ to add a new track'
      operationId: addTrack
      parameters:
        -
          name: dj_id
          in: path
          description: 'DJ ID'
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                name:
                  type: string
                  example: 'Track Name'
                artist:
                  type: string
                  example: 'Artist Name'
                duration:
                  type: string
                  example: '3:45'
              type: object
      responses:
        200:
          description: 'Successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Track'
        400:
          description: 'Bad Request'
        401:
          description: Unauthorized
        404:
          description: 'DJ not found'
      security:
        -
          telegramAuth: []
  '/dj/{dj_id}/tracks':
    get:
      tags:
        - DJ
      summary: 'Get tracks by DJ ID'
      description: 'Returns the list of tracks associated with a specific DJ'
      operationId: getDJTracks
      parameters:
        -
          name: dj_id
          in: path
          description: 'DJ ID'
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 'Successful operation'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Track'
        404:
          description: 'DJ not found'
        401:
          description: Unauthorized
      security:
        -
          telegramAuth: []
  '/dj/{dj_id}/track/{track_id}/price':
    patch:
      tags:
        - DJ
      summary: 'Update track price'
      description: 'Allows a DJ to update the price of a track'
      operationId: updateTrackPrice
      parameters:
        -
          name: dj_id
          in: path
          description: 'DJ ID'
          required: true
          schema:
            type: integer
        -
          name: track_id
          in: path
          description: 'Track ID'
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                price:
                  type: number
                  format: float
                  example: 19.99
              type: object
      responses:
        200:
          description: 'Successful operation'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  price:
                    type: number
                    format: float
                    example: 19.99
                type: object
        400:
          description: 'Bad Request'
        401:
          description: Unauthorized
        404:
          description: 'DJ or Track not found'
      security:
        -
          telegramAuth: []
  '/dj/{dj_id}/statistics':
    get:
      tags:
        - DJ
      summary: 'Get statistics for a DJ'
      description: 'Retrieve various statistics for a DJ including total orders, total income, average price, etc.'
      operationId: 3dbb0c27547adac2bd39534c9122a09f
      parameters:
        -
          name: dj_id
          in: path
          description: 'ID of the DJ'
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 'Successful response'
          content:
            application/json:
              schema:
                properties:
                  total_orders:
                    type: integer
                    example: 5
                  total_income:
                    type: number
                    format: float
                    example: 150.75
                  average_price:
                    type: number
                    format: float
                    example: 30.15
                  min_price:
                    type: number
                    format: float
                    example: 20
                  max_price:
                    type: number
                    format: float
                    example: 50
                  most_popular_tracks:
                    properties:
                      track_id:
                        type: integer
                        example: 1
                      count:
                        type: integer
                        example: 3
                    type: object
                type: object
        404:
          description: 'DJ not found'
          content:
            application/json:
              schema:
                properties:
                  error:
                    type: string
                    example: 'DJ not found'
                type: object
...
components:
  schemas:
...
    DJ:
      properties:
        id:
          type: integer
          example: 1
        user_id:
          type: integer
          example: 1
        stage_name:
          type: string
          example: 'DJ Example'
        city:
          type: string
          example: 'New York'
        payment_details:
          type: string
          example: 'Bank details or any payment information'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      type: object
    Track:
      title: Track
      required:
        - name
        - artist
        - duration
      properties:
        id:
          description: 'Track ID'
          type: integer
          example: 1
        name:
          description: 'Track name'
          type: string
          example: 'Track Name'
        artist:
          description: 'Artist name'
          type: string
          example: 'Artist Name'
        duration:
          description: 'Track duration'
          type: string
          example: '3:45'
        created_at:
          description: 'Creation timestamp'
          type: string
          format: date-time
          example: '2024-07-01T00:00:00Z'
        updated_at:
          description: 'Last update timestamp'
          type: string
          format: date-time
          example: '2024-07-01T00:00:00Z'
      type: object
...
  securitySchemes:
    telegramAuth:
      type: apiKey
      in: header
      name: Telegram-Init-Data
      description: 'Value of TWA.initData() from Telegram Web App.'
...
security:
  -
    telegramAuth: []